requirements.txt
annotated-types==0.7.0
anyio==4.9.0
beautifulsoup4==4.13.4
bs4==0.0.2
certifi==2025.6.15
charset-normalizer==3.4.2
distro==1.9.0
dotenv==0.9.9
greenlet==3.2.3
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
jiter==0.10.0
openai==1.91.0
playwright==1.53.0
pydantic==2.11.7
pydantic_core==2.33.2
pyee==13.0.0
python-dateutil==2.9.0.post0
python-dotenv==0.21.1
python-substack==0.1.15
PyYAML==6.0.2
requests==2.32.4
six==1.17.0
sniffio==1.3.1
soupsieve==2.7
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.14.0
urllib3==2.5.0
